Update my_map to be in one file
