+++ README.md
@@ -56,20 +56,18 @@ you will need to implement
 along with their associated specs,
 as outlined in the project instructions.
 
-| Method                              | Spec File                         | Notes                                                                 |
-| ----------------------------------- | --------------------------------- | --------------------------------------------------------------------- |
-| `Array#my_each`                     | spec/my_each_spec.rb              | You will define my_each on the Array class in `lib/my_enumerables.rb` |
-| `Enumerable#my_each_with_index`     | spec/my_each_with_index_spec.rb   |                                                                       |
-| `Enumerable#my_select`              | spec/my_select_spec.rb            |                                                                       |
-| `Enumerable#my_all?`                | spec/my_all_spec.rb               |                                                                       |
-| `Enumerable#my_any?`                | spec/my_any_spec.rb               |                                                                       |
-| `Enumerable#my_none?`               | spec/my_none_spec.rb              |                                                                       |
-| `Enumerable#my_count`               | spec/my_count_spec.rb             |                                                                       |
-| `Enumerable#my_map { block }`       | spec/my_map_spec.rb               | This method only needs to take a block                                |
-| `Enumerable#my_inject`              | spec/my_inject_spec.rb            |                                                                       |
-| `multiply_els`                      | spec/multiply_els_spec.rb         | You will need to complete this function in `lib/multiply_els.rb`      |
-| `Enumerable#my_map(proc)`           | spec/my_map_proc_spec.rb          | The my_map method, but takes a proc instead of a block                |
-| `Enumerable#my_map(proc) { block }` | spec/my_map_proc_or_block_spec.rb | the my_map method, but takes a proc and/or a block                    |
+| Method                          | Spec File                       | Notes                                                                 |
+| ------------------------------- | ------------------------------- | --------------------------------------------------------------------- |
+| `Array#my_each`                 | spec/my_each_spec.rb            | You will define my_each on the Array class in `lib/my_enumerables.rb` |
+| `Enumerable#my_each_with_index` | spec/my_each_with_index_spec.rb |                                                                       |
+| `Enumerable#my_select`          | spec/my_select_spec.rb          |                                                                       |
+| `Enumerable#my_all?`            | spec/my_all_spec.rb             |                                                                       |
+| `Enumerable#my_any?`            | spec/my_any_spec.rb             |                                                                       |
+| `Enumerable#my_none?`           | spec/my_none_spec.rb            |                                                                       |
+| `Enumerable#my_count`           | spec/my_count_spec.rb           |                                                                       |
+| `Enumerable#my_map`             | spec/my_map_spec.rb             |                                                                       |
+| `Enumerable#my_inject`          | spec/my_inject_spec.rb          |                                                                       |
+| `multiply_els`                  | spec/multiply_els_spec.rb       | You will need to complete this function in `lib/multiply_els.rb`      |
 
 ### Example
